fix: backends now fully use size_t instead of int

+++ backends/cpu/include/cpu_backend/CPUBackend.h
@@ -177,8 +177,8 @@ public:
     void complexDivision(const ComplexData& a, const ComplexData& b, ComplexData& result, real_t epsilon) const override;
     void division(const RealData& a, const RealData& b, RealData& result, real_t epsilon) const override;
     void complexAddition(const ComplexData& a, const ComplexData& b, ComplexData& result) const override;
-    void complexAddition(complex_t** data, ComplexData& sum, int nImages, size_t imageVolume) const override;
-    void sumToOne(real_t** data, int nImages, size_t imageVolume) const override;
+    void complexAddition(complex_t** data, ComplexData& sum, size_t nImages, size_t imageVolume) const override;
+    void sumToOne(real_t** data, size_t nImages, size_t imageVolume) const override;
     void sum(const ComplexData& data, complex_t* result) const override;
     void meanSquareError(const ComplexData& a, const ComplexData& b, real_t* result) const override;
     void scalarMultiplication(const ComplexData& a, complex_t scalar, ComplexData& result) const override;
